20110120_0930 commit: add staffMemberDetail view

+++ src/main/java/org/mcguppy/eventplaner/business/dispomgmt/boundary/CrudBaseService.java
@@ -29,7 +29,7 @@ public abstract class CrudBaseService<T> {
         getEntityManager().remove(getEntityManager().merge(entity));
     }
 
-    public T find(Long id) {
+    public T find(Integer id) {
         return getEntityManager().find(entityClass, id);
     }
